REFACTORED CODEBASE + instable not working very well
