+++ client/services/auth/GetOTPcode.go
@@ -1,4 +1,4 @@
-package services
+package auth
 
 import (
 	"bufio"
@@ -6,6 +6,7 @@ import (
 	"os"
 	"strings"
 
+	inoutput "github.com/Gustavo-DCosta/EchoMail/client/services/io"
 	"github.com/fatih/color"
 )
 
@@ -17,7 +18,7 @@ func GetSMScode() string {
 
 	input, err := reader.ReadString('\n')
 	if err != nil {
-		Check(err)
+		inoutput.Check(err)
 		fmt.Println("Error sanitizing the input", err)
 		return ""
 	}
@@ -25,7 +26,7 @@ func GetSMScode() string {
 	if input == "" {
 		stdOutOTPcode.Print("The code seems empty Whiskers McGraw doesn't aprove it\n")
 		stdOutOTPcode.Print("Please wirte it again:	")
-		InfoLogs("There was a problem to receive the OTP from user input")
+		inoutput.InfoLogs("There was a problem to receive the OTP from user input")
 	}
 
 	OTPcode := strings.TrimSpace(input)
