--- client/services/core/clear.go
@@ -1,9 +1,0 @@
-package services
-
-import (
-	"fmt"
-)
-
-func ClearUI() {
-	fmt.Print("\033[2J\033[H")
-}
