+++ client/services/network/connService.go
@@ -1,4 +1,4 @@
-package services
+package network
 
 import (
 	"bytes"
@@ -9,12 +9,13 @@ import (
 	"os"
 
 	"github.com/Gustavo-DCosta/EchoMail/client/model"
+	inoutput "github.com/Gustavo-DCosta/EchoMail/client/services/io"
 )
 
 func SendConnCredentials(phoneNumber, emaillAddress string, newUser bool) (string, error) {
 	url := os.Getenv("ServerConnUrl")
 	if url == "" {
-		InfoLogs("Couldn't get the url from the env file")
+		inoutput.InfoLogs("Couldn't get the url from the env file")
 		return "", fmt.Errorf("environment variable ServerConnUrl is not set")
 	}
 
@@ -27,14 +28,14 @@ func SendConnCredentials(phoneNumber, emaillAddress string, newUser bool) (strin
 	reqPayload, err := json.Marshal(payloadStruct)
 	if err != nil {
 		fmt.Println("Error marsheling the payload structure | ERROR: ", err)
-		Check(err)
+		inoutput.Check(err)
 		return "", err
 	}
 
 	req, err := http.NewRequest("POST", url, bytes.NewBuffer(reqPayload))
 	if err != nil {
 		fmt.Println("Error making a new request | ERROR: ", err)
-		Check(err)
+		inoutput.Check(err)
 		return "", err
 	}
 
@@ -43,7 +44,7 @@ func SendConnCredentials(phoneNumber, emaillAddress string, newUser bool) (strin
 	client := &http.Client{}
 	resp, err := client.Do(req)
 	if err != nil {
-		Check(err)
+		inoutput.Check(err)
 		fmt.Println("Couldn't send the http request | ERROR :", err)
 		return "", nil
 	}
@@ -57,14 +58,14 @@ func SendConnCredentials(phoneNumber, emaillAddress string, newUser bool) (strin
 	var serverResponse model.UUIDResponse
 	bodyBytes, err := io.ReadAll(resp.Body)
 	if err != nil {
-		Check(err)
+		inoutput.Check(err)
 		fmt.Println("Couldn't read the response body | ERROR:", err)
 		return "", err
 	}
 
 	if err := json.Unmarshal(bodyBytes, &serverResponse); err != nil {
 		fmt.Println("Couldn't unmarshal the uuid response | ERROR:", err)
-		Check(err)
+		inoutput.Check(err)
 		return "", err
 	}
